if latest is not valid set point in time to mined past
