+++ src/eth/storage/mod.rs
@@ -6,9 +6,7 @@ pub use permanent::PermanentStorageConfig;
 pub use permanent::RocksPermanentStorage;
 pub use stratus_storage::StratusStorage;
 pub use temporary::InMemoryTemporaryStorage;
-pub use temporary::ReadKind;
 pub use temporary::TemporaryStorageConfig;
-pub use temporary::TxCount;
 
 mod cache;
 pub mod permanent;
@@ -23,6 +21,7 @@ use display_json::DebugAsJson;
 
 use crate::eth::primitives::Account;
 use crate::eth::primitives::Address;
+use crate::eth::primitives::BlockNumber;
 use crate::eth::primitives::Slot;
 use crate::eth::primitives::SlotIndex;
 use crate::eth::primitives::StratusError;
@@ -77,4 +76,57 @@ impl StorageConfig {
 
         Ok(Arc::new(storage))
     }
+}
+
+
+#[derive(Clone, Copy, PartialEq, Debug, serde::Serialize, serde::Deserialize, fake::Dummy, Eq)]
+pub enum TxCount {
+    Full,
+    Partial(u64),
+}
+
+impl From<u64> for TxCount {
+    fn from(value: u64) -> Self {
+        TxCount::Partial(value)
+    }
+}
+
+impl Default for TxCount {
+    fn default() -> Self {
+        TxCount::Partial(0)
+    }
+}
+
+impl std::ops::AddAssign<u64> for TxCount {
+    fn add_assign(&mut self, rhs: u64) {
+        match self {
+            TxCount::Full => {}                       // If it's Full, keep it Full
+            TxCount::Partial(count) => *count += rhs, // If it's Partial, increment the counter
+        }
+    }
+}
+
+impl Ord for TxCount {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        match (self, other) {
+            (TxCount::Full, TxCount::Full) => std::cmp::Ordering::Equal,
+            (TxCount::Full, TxCount::Partial(_)) => std::cmp::Ordering::Greater,
+            (TxCount::Partial(_), TxCount::Full) => std::cmp::Ordering::Less,
+            (TxCount::Partial(a), TxCount::Partial(b)) => a.cmp(b),
+        }
+    }
+}
+
+impl PartialOrd for TxCount {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+#[derive(Clone, Copy, serde::Serialize, serde::Deserialize, PartialEq, Default, fake::Dummy)]
+pub enum ReadKind {
+    Call((BlockNumber, TxCount)),
+    #[default]
+    Transaction,
+    RPC,
 }
